feat: isolated price logic into a different class
